refactor: types related to prisma switched to interfaces/prisma
