+++ src/routes/graphql/interfaces/prisma/prismaContext.ts
@@ -1,0 +1,7 @@
+import { PrismaClient } from "@prisma/client";
+import { DataLoaderMapT } from "../../index.js";
+
+export type IPrismaContext = {
+  prisma: PrismaClient;
+  dataloaders: DataLoaderMapT;
+};
